Enabled structured logging.
Added a wellknown structured logging lib [sirupsen/logrus](github.com/sirupsen/logrus) to allow for structured logging, this can be enabled with logging_mode env var set to `json`
Sswapped to using a newline scanner for reading from the stderr and stdout pipes of the wrapped functions.  This fixes cases where the previously limited 256 bytes caused splits in log output from the inner function.
The HTTP Watchdog was slightly changed to use `.Wait` on the process instead of calling Fatal when failing to read from stdout.